create testFixtures for actions; fix seqCons for tab
