+++ road-runner/actions/src/test/kotlin/com/acmerobotics/roadrunner/actions/ActionsTest.kt
@@ -1,10 +1,7 @@
-package com.acmerobotics.roadrunner
+package com.acmerobotics.roadrunner.actions
 
 import com.acmerobotics.dashboard.canvas.Canvas
 import com.acmerobotics.dashboard.telemetry.TelemetryPacket
-import com.acmerobotics.roadrunner.actions.Action
-import com.acmerobotics.roadrunner.actions.ParallelAction
-import com.acmerobotics.roadrunner.actions.SequentialAction
 import org.junit.jupiter.api.Test
 import kotlin.test.assertEquals
